web: add seat positions, deepcopy Player with lodash
Because the player's client/player information is saved in a useState
variable, we need to deepcopy the Player object to ensure that invalid
player actions are not reflected in the front-end. Any changes made to
any publicly visible information is first validated and sent back to all
the clients by the server.

Also:
 - Fix player action keyboard shortcut bugs
 - Change vacant seat display
 - Clear bet input after pressing `raise`
 - Ensure cards are not rendered for vacant seats/mid-round additions
 - cleanup properly on ELIMINATED & RESET (wip)
 - Allow client to pause logo spin
 - Log the server responses in a large red font (for debugging)