implement reconnecting feature on server and web
Previously, when a websocket would close (e.g. on a shaky connection)
there would be no attempt to reconnect, and the frontend would give no
indication that the websocket was closed. I've added a private ID to
each client so that they can reconnect with it. Implemented on web
frontend, not yet on CLI frontend.