check if room exists before attempting reconnect
When the last client in a room is disconnected for more than 1 minute,
the server automatically removes the room. The WebSocket
constructor throws an error when the server returns a 404 before the WS
upgrade, and the 404 error is lost; causing the error div to be rendered
with an unspecified error. Because of this, we simply call the checkRoom
API and render the appropriate error div.

Also:
main.go
 - Check for the 404 response code on a Dial() err and print the
   appropriate error message
net/net.go:
 - When a player is disconnected and the server is waiting for their
   reconnect, the other players can continue until it's the disconnected
   player(s) turn. The disconnected players are still in CurPlayers() and
   ActivePlayers(), and will be still be sent private info like during
   sendDeals(); therefore we check isDisconnected in
   NetData.{Send,SendTo}() to make sure we don't panic() when their conn
   field is nil
net/room.go:
 - room.connClientMap[room.getPlayerConn(player)] was left over from
   older code and was replaced with the newer and faster getPlayerClient()