temporary fix to reconnect bug
Previously, the client cleanup AfterFunc would sit around after a player
reconnected. If a player reconnects multiple times in <60sec time spans
the chance of one of these firing while waiting for a newer reconnect
becomes more and more likely; causing the cleanup to happen when it
shouldn't. Instead, keep check of timer AfterFunc returns and stop the
timer where appropriate. I've tried to avoid adding a mutex to the
Player struct but it's clear that I need to add it later, hence the
temporary label.