properly escape room names
Reserved characters weren't being escaped properly in the web frontend
which caused chars like # to be treated as special parts of the URL. For
convenience, the server will now return an escaped URL in the checkRoom()
JSON response as well as the roomName. The web frontend escapes the roomName
with encodeURIComponent() as appropriate; pass a URLObject to router.{push,replace}
instead of a string to get automatic percent encoding of the room names.